Separate SQL query building from imported.ex
